WorldSession Relocation and Refactoring
The WorldSession class has been moved to Game/World with separate header and implementation sections. The includes have been fixed, and CMakeLists.txt has been updated to include new files and dependencies (including the addition of fireland_crypto). The async includes have been adjusted. The GetCharacterByGuid method now uses uint64_t to support 64-bit GUIDs. This refactoring has no functional impact and improves the project's modularity and clarity.

+++ src/common/Database/include/Database/Char/CharWrapper.h
@@ -33,7 +33,7 @@ namespace Fireland::Database::Char
 
         // -- Characters operations
         Fireland::Utils::Async::async<bool> IsNameAvailable(std::string_view name) noexcept;
-        Fireland::Utils::Async::async<std::optional<characters>> GetCharacterByGuid(uint32_t guid) noexcept;
+        Fireland::Utils::Async::async<std::optional<characters>> GetCharacterByGuid(uint64_t guid) noexcept;
         Fireland::Utils::Async::async<std::vector<characters>> GetCharactersForAccount(uint32_t accountid) noexcept;
         Fireland::Utils::Async::async<std::optional<characters>> CreateCharacter(characters c) noexcept;
         Fireland::Utils::Async::async<bool> DeleteCharacter(uint64_t guid, uint32_t accountid) noexcept;
